backend/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.38.0
python-multipart==0.0.6
Pillow==10.1.0
cryptography==41.0.7
pydantic==2.5.0
python-dotenv==1.2.1